annotation_app/requirements.txt
albumentations==1.2.1
CairoSVG==2.4.2
dash==2.6.1
dash_bootstrap_components==1.2.1
dash_bootstrap_templates==1.0.6
dash_core_components==2.0.0
dash_html_components==2.0.0
dash_table==5.0.0
Flask==1.1.2
Flask_Assets==2.0
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
joblib==1.0.1
matplotlib==3.2.2
numpy==1.18.4
pandas==1.1.5
pandas_datareader==0.10.0
Pillow==9.3.0
plotly==5.10.0
python-dotenv==0.21.0
scikit_image==0.16.2
scikit_learn==1.1.3
segmentation_models_pytorch==0.3.0
skimage==0.0
SQLAlchemy==1.4.22
torch==1.10.2
torchvision==0.11.3
Werkzeug==2.0.3
WTForms==3.0.0
